+++ src/index.ts
@@ -8,6 +8,7 @@ export * from "./components/forms/FormColorInput.js";
 export * from "./components/forms/FormDateInput.js";
 export * from "./components/forms/FormEmailInput.js";
 export * from "./components/forms/FormFieldset.js";
+export * from "./components/forms/FormFileInput.js";
 export * from "./components/forms/FormHiddenInput.js";
 export * from "./components/forms/FormLabel.js";
 export * from "./components/forms/FormNoticeContainer.js";